Add recipes to profile view
Modify search results on homepage to not have a delete button.

Recipes owned by user are displayed on profile page when logged in.

Recipe page template moved from `home.pug` to `app.pug` to allow for
reuse between home page and profile page for displaying recipes.

Modified `login.js` to return user-owned recipes to display when log in
is successful.

Closes #96