patch: fix pantry add save status being cut off
Pantry add save status was cut off by the submit button on mobile because
the column width was `1` to keep the save status and the buttons
centered and unmoving while the save status changed. Column width of `1`
however did not appear nicely on mobile and displaying the save status
caused the submit button to overlap the bleeding text.

Save status was moved to a separate row to allow it to expand as needed
without getting cut off.